adding new client feature + service added
