+++ angular-src/src/app/components/calls-meeting-piechart/calls-meeting-piechart.component.html
@@ -1,0 +1,9 @@
+<p>
+  calls-meeting-piechart works!
+</p>
+
+<div class="pie-chart-container">	
+	<app-loading-spinner [hide]="!isLoading"></app-loading-spinner>
+	<h4>{{ 'Calls - Meeting (Count)' }}</h4>
+	<div id="piechart"></div>
+</div>
