+++ angular-src/src/app/components/calls-meeting-piechart/calls-meeting-piechart.component.ts
@@ -1,0 +1,20 @@
+import { Component, OnInit } from '@angular/core';
+import { ReportService } from '../../services/report.service';
+import * as ChartConfig from '../../config/column-chart-config';
+
+const Highcharts = require("highcharts");
+
+
+@Component({
+  selector: 'app-calls-meeting-piechart',
+  templateUrl: './calls-meeting-piechart.component.html',
+  styleUrls: ['./calls-meeting-piechart.component.css']
+})
+export class CallsMeetingPiechartComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
