+++ angular-src/src/app/components/login/login.component.css
@@ -1,0 +1,3 @@
+a#forgot_password_link {
+    color: #0070d2;
+}
