+++ angular-src/src/app/components/login/login.component.html
@@ -10,5 +10,10 @@
     <label>Password</label>
     <input type="password" class="form-control" [(ngModel)]="password" name="password">
   </div>
+  
+<!--    <div class="form-group">
+  		<a id="forgot_password_link" class="fl small" href="/secur/forgotpassword.jsp?locale=us&amp;lqs=">Forgot Your Password?</a>
+	</div>
+ -->
   <input type="submit" class="btn btn-primary" value="Login">
 </form>
